backend/requirements.txt
asgiref==3.11.1
contourpy==1.3.3
cycler==0.12.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
filelock==3.25.2
fonttools==4.61.1
fsspec==2026.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
patsy==1.0.2
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sqlparse==0.5.5
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0+cpu
typing_extensions==4.15.0
